finishing the delete functionality

+++ src/components/Product.tsx
@@ -1,6 +1,6 @@
-import React from 'react'
 import { useDispatch } from 'react-redux'
 import { Link } from 'react-router-dom'
+import Swal from 'sweetalert2'
 import { deleteProductAction } from '../actions/productActions'
 
 const Product = ({ product }) => {
@@ -9,8 +9,21 @@ const Product = ({ product }) => {
     const dispatch = useDispatch()
 
     const deleteProductConfirmation = id => {
-        dispatch(deleteProductAction(id))
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteProductAction(id))
+            }
+        })
     }
+
     return (
         <tr>
             <td>{name}</td>
